+++ packages/solidity-devops/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4](https://github.com/synapsecns/sanguine/compare/@synapsecns/solidity-devops@0.1.3...@synapsecns/solidity-devops@0.1.4) (2024-03-04)
+
+**Note:** Version bump only for package @synapsecns/solidity-devops
+
+
+
+
+
 ## [0.1.3](https://github.com/synapsecns/sanguine/compare/@synapsecns/solidity-devops@0.1.2...@synapsecns/solidity-devops@0.1.3) (2024-03-01)
 
 **Note:** Version bump only for package @synapsecns/solidity-devops